Gets a couple of download classes to compile on linux. Updates the views build so that we only include the gtk classes we need.
BUG=none
TEST=none

Review URL: http://codereview.chromium.org/115788

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@16899 0039d316-1c4b-4281-b951-d872f2087c98

+++ chrome/browser/views/download_shelf_view.cc
@@ -23,7 +23,7 @@
 
 // Max number of download views we'll contain. Any time a view is added and
 // we already have this many download views, one is removed.
-static const int kMaxDownloadViews = 15;
+static const size_t kMaxDownloadViews = 15;
 
 // Padding from left edge and first download view.
 static const int kLeftPadding = 2;
